Refactoring of Employee Controller, Service and Repo and change Employee id type to int

+++ LibraryManager/Migrations/LibraryManagerContextModelSnapshot.cs
@@ -40,9 +40,9 @@ namespace LibraryManager.Migrations
 
             modelBuilder.Entity("LibraryManager.Models.Employee", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("FirstName")
